perf(zerocheck): MLE interpolate directly in eval kernel and use shmem (#330)
Previous zerocheck/logup MLE kernels first did linear interpolation and
then did eval directly on the matrices. Given it's just linear
interpolation, it seems better to just interpolate on-demand. (edit:
benchmarking shows using buffer vars is worse, and interpolation into a
new buffer may indeed be better.)

More importantly, I removed the TASK_SIZE limit since for the later MLE
rounds I think we shouldn't be memory bound so we should maximize
parallelism. Also I did the shared memory trick to reduce the sum within
a block prior to the second block reduce kernel.

+++ crates/cuda-backend-v2/cuda/src/logup_zerocheck/zerocheck_round0.cu
@@ -250,7 +250,7 @@ constexpr uint32_t MAX_THREADS = 256;
 
 // (Not a launcher) Utility function to calculate required size of temp sum buffer.
 // Required length of *temp_sum_buffer in FpExt elements
-extern "C" uint32_t _zerocheck_r0_temp_sums_buffer_size(
+extern "C" size_t _zerocheck_r0_temp_sums_buffer_size(
     uint32_t buffer_size,
     uint32_t large_domain,
     uint32_t num_x,
@@ -261,7 +261,7 @@ extern "C" uint32_t _zerocheck_r0_temp_sums_buffer_size(
     );
 }
 
-extern "C" uint32_t _zerocheck_r0_intermediates_buffer_size(
+extern "C" size_t _zerocheck_r0_intermediates_buffer_size(
     uint32_t buffer_size,
     uint32_t large_domain,
     uint32_t num_x,
